+++ src/Block/Parser/SetExtHeadingParser.php
@@ -47,7 +47,7 @@ class SetExtHeadingParser implements BlockParserInterface
         $level = $match[0][0] === '=' ? 1 : 2;
         $strings = $context->getContainer()->getStrings();
 
-        $strings = $this->resolveReferenceLinkDefinitions($strings, $cursor->getEncoding(), $context->getReferenceParser());
+        $strings = $this->resolveReferenceLinkDefinitions($strings, $context->getReferenceParser());
         if (empty($strings)) {
             return false;
         }
@@ -63,15 +63,14 @@ class SetExtHeadingParser implements BlockParserInterface
      * @see https://github.com/commonmark/commonmark.js/commit/993bbe335931af847460effa99b2411eb643577d
      *
      * @param string[]        $strings
-     * @param string          $encoding
      * @param ReferenceParser $referenceParser
      *
      * @return string[]
      */
-    private function resolveReferenceLinkDefinitions(array $strings, string $encoding, ReferenceParser $referenceParser): array
+    private function resolveReferenceLinkDefinitions(array $strings, ReferenceParser $referenceParser): array
     {
         foreach ($strings as &$string) {
-            $cursor = new Cursor($string, $encoding);
+            $cursor = new Cursor($string);
             while ($cursor->getCharacter() === '[' && $referenceParser->parse($cursor)) {
                 $string = $cursor->getRemainder();
             }